requirements.txt
Django==5.0.6
whitenoise==6.4.0
gunicorn==21.2.0
Pillow>=10.3.0
psycopg2-binary
